Extract header components
