+++ src/Header/types.ts
@@ -1,0 +1,6 @@
+export interface HeaderProps {
+  sections: ReadonlyArray<{
+    title: string;
+    url: string;
+  }>;
+}
